SetCover exercise
A program that finds the smallest subset of sets, which contains all elements from a given sequence.
In the Set Cover Problem, we are given two sets - a set of sets (we’ll call it sets) and a universe (a sequence).
The sets contain all elements from the universe and no others, however, some elements are repeated. The task is
to find the smallest subset of sets that contains all elements in the universe